Attacks now fully work
- Added a HitCollider script for every object that holds the player's attack collider, this script is responsible for letting the AttackColliders script know that a collider was triggered
- Now resets after hitting the final hit
- Now damages an IDamageable object by the player's damage, with the last hit dealing 3x damage

+++ Assets/_Scripts/Logic/Interfaces/IDamageable.cs
@@ -1,5 +1,5 @@
 
     public interface IDamageable
     {
-        public void Damage(int damageAmount);
+        public void Damage(float damageAmount);
     }
